requirements.txt
giteapy>=1.0.8
PyGithub>=1.55